refactor(runner): carve environment-setup.ts out of acquireEnvironment
Lift the setup prefix of acquireEnvironment (logger/timing, ownership check, mount
signing, durable cwd, buildRunPlan, daemon env, pi extension env, binary resolve,
initial SessionEnvironment construction) into prepareEnvironmentSetup, which returns
a bundle that acquireEnvironment destructures. Statement order and side effects
unchanged; typecheck enforces the bundle is complete. 1158 runner unit tests pass.
