+++ services/runner/src/engines/sandbox_agent/environment.ts
@@ -45,8 +45,6 @@ import {
   resolveRunSessionId,
   type AgentRunRequest,
 } from "../../protocol.ts";
-import { type ClientToolOutcome } from "../../responder.ts";
-import type { ClientToolRelay } from "../../tools/client-tool-relay.ts";
 import { advertisedToolSpecs } from "../../tools/public-spec.ts";
 import { createAcpFetch } from "./acp-fetch.ts";
 import {
@@ -54,8 +52,6 @@ import {
   assertRequiredCapabilities,
   probeCapabilities,
 } from "./capabilities.ts";
-import { createToolCallCorrelationIndex } from "./client-tools.ts";
-import { buildDaemonEnv, resolveDaemonBinary } from "./daemon.ts";
 import { DaytonaReconnectTerminalError } from "./daytona-provider.ts";
 import {
   createCookieFetch,
@@ -77,15 +73,10 @@ import {
   type MountCredentials,
 } from "./mount.ts";
 import {
-  buildPiExtensionEnv,
-  configurePiSessionWorkspace,
-  configurePiSkillSnapshot,
   PI_PERMISSION_EXTENSION_UNAVAILABLE_MESSAGE,
   prepareLocalPiAssets,
-  resolvePiSkillSnapshot,
   uploadSystemPromptToSandbox,
   writeSystemPromptLocal,
-  writeOtlpAuthFile,
 } from "./pi-assets.ts";
 import {
   AGENT_MOUNT_ENV_VAR,
@@ -94,7 +85,6 @@ import {
   linkAgentFilesRemote,
   seedAgentReadme,
   seedAgentReadmeRemote,
-  signAgentMountCredentials,
 } from "./agent-mount.ts";
 import {
   AGENT_MOUNT_SYSTEM_PROMPT_SEGMENT,
@@ -106,9 +96,7 @@ import {
   routePermissionRequestToActiveTurn,
   routeSessionEventToActiveTurn,
 } from "./session-events.ts";
-import { loadRunnerConfig } from "../../config/runner-config.ts";
 import { buildSandboxProvider } from "./provider.ts";
-import { buildRunPlan } from "./run-plan.ts";
 import {
   clearSandboxPointer,
   readStoredSandboxPointer,
@@ -120,24 +108,17 @@ import type {
   SessionEnvironment,
 } from "./runtime-contracts.ts";
 import {
-  applyClaudeConnectionEnv,
   containsTransportEndpointDisconnected,
-  defaultResolveLocalRunnerOwner,
   isTransportEndpointDisconnected,
-  modelResolutionStrict,
   runCredential,
 } from "./runtime-policy.ts";
 import { hydrateHarnessSessionFromDurable } from "./session-continuity-durable.ts";
 import {
-  assertLocalRunnerOwnership,
   eligibleAgentSessionId,
   nextTurnIndex,
   sessionContinuityStore,
 } from "./session-continuity.ts";
-import {
-  projectScopeFor,
-  resolvesToLocalProvider,
-} from "./session-identity.ts";
+import { projectScopeFor } from "./session-identity.ts";
 import {
   teardownDisposition,
   type TeardownReason,
@@ -147,6 +128,7 @@ import {
   type ToolMcpAssets,
 } from "./tool-mcp-assets.ts";
 import { prepareWorkspace } from "./workspace.ts";
+import { prepareEnvironmentSetup } from "./environment-setup.ts";
 
 function log(message: string): void {
   process.stderr.write(`[sandbox-agent] ${message}\n`);
@@ -273,266 +255,31 @@ export async function acquireEnvironment(
   signal?: AbortSignal,
   presignedMount?: MountCredentials | null,
 ): Promise<AcquireEnvironmentResult> {
-  const logger = deps.log ?? log;
-  const acquireStartedAt = Date.now();
-  const timingLog = (stage: string, startedAt: number, fields = ""): void => {
-    const sandboxId = environment?.sandbox?.sandboxId ?? "-";
-    const sessionId =
-      environment?.sessionId ?? request.sessionId?.trim() ?? "-";
-    logger(
-      `[timing] stage=${stage} ms=${Math.round(Date.now() - startedAt)} sandbox=${sandboxId} session=${sessionId}${fields}`,
-    );
-  };
-
-  // Local multi-runner fails loudly. Session-owned + local-sandbox only (a non-session run
-  // has no cross-replica identity to protect, and a remote sandbox has no runner-local pooled
-  // state to protect it FROM). The resolver claims the `owner` affinity key and reads the actual
-  // owner back; a KNOWN different owner throws (never a silent wrong-host cold start).
-  const continuitySessionForOwnership = request.sessionId?.trim();
-  if (
-    continuitySessionForOwnership &&
-    resolvesToLocalProvider(request.sandbox)
-  ) {
-    const { replicaId, ownerReplicaId } = await (
-      deps.resolveLocalRunnerOwner ?? defaultResolveLocalRunnerOwner
-    )(continuitySessionForOwnership, runCredential(request));
-    try {
-      assertLocalRunnerOwnership(
-        continuitySessionForOwnership,
-        replicaId,
-        ownerReplicaId,
-      );
-    } catch (err) {
-      return { ok: false, error: conciseError(err, request.harness ?? "") };
-    }
-  }
-
-  // Sign BEFORE buildRunPlan so the prefix is available for the durable cwd derivation.
-  // Inputs (sessionId, apiBase, credential) are independent of the plan. Best-effort: null on
-  // failure leaves durableCwd undefined and buildRunPlan falls back to the ephemeral path.
-  const sessionForMount = request.sessionId?.trim();
-  const runCred = runCredential(request);
-  const signMount =
-    deps.signSessionMountCredentials ?? signSessionMountCredentials;
-  let mountCreds: MountCredentials | null =
-    presignedMount !== undefined
-      ? presignedMount
-      : sessionForMount && runCred
-        ? await signMount(sessionForMount, {
-            apiBase: apiBase(),
-            authorization: runCred,
-            log: logger,
-          })
-        : null;
-  // A session-owned run expects a durable session cwd mount. When signing returns nothing the run
-  // still proceeds on an ephemeral cwd (behavior unchanged, RSH-11); emit one structured warning
-  // keyed by mount kind so durable-to-ephemeral degradation is measurable, not silent.
-  if (sessionForMount && !mountCreds) {
-    logger(
-      `mount degraded kind=session_cwd cause=sign_returned_no_mount session=${sessionForMount}`,
-    );
-  }
-
-  const artifactId = request.runContext?.workflow?.artifact?.id?.trim();
-  const signAgentMount =
-    deps.signAgentMountCredentials ?? signAgentMountCredentials;
-  const agentMountCreds: MountCredentials | null =
-    artifactId && runCred
-      ? await signAgentMount(artifactId, {
-          apiBase: apiBase(),
-          authorization: runCred,
-          log: logger,
-        })
-      : null;
-  // A workflow-artifact run expects an agent mount; same structured degrade signal when unsigned.
-  if (artifactId && !agentMountCreds) {
-    logger(
-      `mount degraded kind=agent_mount cause=sign_returned_no_mount artifact=${artifactId}`,
-    );
-  }
-  // Derive the durable cwd from the sign prefix (one source of truth, both providers).
-  // local: /tmp/agenta/<prefix>  —  daytona: /home/sandbox/agenta/<prefix>
-  // <prefix> is already "mounts/<project_id>/<mount_id>", so no extra slug is needed.
-  let durableCwd: string | undefined;
-  if (mountCreds?.prefix) {
-    const isDaytonaReq =
-      (request.sandbox ?? loadRunnerConfig().providers.default) === "daytona";
-    durableCwd = isDaytonaReq
-      ? `/home/sandbox/agenta/${mountCreds.prefix}`
-      : `/tmp/agenta/${mountCreds.prefix}`;
-  }
-
-  const planResult = buildRunPlan(request, {
-    sandboxProvider: deps.sandboxProvider,
-    createLocalCwd: deps.createLocalCwd,
-    createDaytonaCwd: deps.createDaytonaCwd,
-    durableCwd,
-    resolveSkillDirs: deps.resolveSkillDirs,
-    log: logger,
-  });
-  if (!planResult.ok) return { ok: false, error: planResult.error };
-  const plan = planResult.plan;
-  const piSkillSnapshot = resolvePiSkillSnapshot(plan);
-  const agentMountDir = agentMountCreds ? agentMountPath(plan.cwd) : undefined;
-
-  // Clear-then-apply (Security rule 5): on a managed run (credentialMode "env") the daemon
-  // inherits NONE of the sidecar's own provider keys, so only the resolved `plan.secrets` are
-  // present and an inherited key for another provider cannot leak. For runtime_provided/none/
-  // un-migrated runs the harness uses its own login, so the inherited keys stay.
-  const clearProviderEnv = plan.credentialMode === "env";
-  const env = (deps.buildDaemonEnv ?? buildDaemonEnv)(plan.acpAgent, {
-    clearProviderEnv,
-    provider: request.provider,
-    deployment: request.deployment,
-  });
-  Object.assign(env, plan.secrets); // apply only the resolved provider keys
-  applyClaudeConnectionEnv(env, request, plan.acpAgent, logger);
-  const piSessionDir = configurePiSessionWorkspace(plan, env);
-  configurePiSkillSnapshot(piSkillSnapshot, env);
-  const strictModel = modelResolutionStrict();
-  // Pi self-instruments locally: propagate the trace context + public tool metadata into Pi
-  // via the Agenta extension. Tool execution always relays back to this runner, which keeps
-  // private specs, scoped env, callback endpoints, and callback auth in memory.
-  // local Pi's OTLP bearer rides a runner-written 0600 file, never a plain env var —
-  // Daytona never receives telemetry env here at all (`!plan.isDaytona` gates it off above).
-  const otlpAuthFilePath =
-    plan.isPi && !plan.isDaytona ? `${plan.relayDir}.otlp-auth` : undefined;
-  const otlpAuthorization =
-    request.telemetry?.exporters?.otlp?.headers?.authorization;
-  if (otlpAuthFilePath && otlpAuthorization) {
-    writeOtlpAuthFile(otlpAuthFilePath, otlpAuthorization, logger);
-  }
-  const piExtEnv = plan.isPi
-    ? buildPiExtensionEnv(request, !plan.isDaytona, {
-        relayDir: plan.relayDir,
-        usageOutPath: plan.usageOutPath,
-        otlpAuthFilePath,
-        builtinGatingActive: plan.builtinGatingActive,
-        builtinGrants: plan.builtinGrants,
-        // The materialized skill names (author + forced `_agenta.*`) so Pi's own agent span
-        // records which skills loaded; local Pi self-instruments, so the runner's sandbox-agent
-        // otel has no span to stamp here.
-        skills: plan.skillDirs.map((s) => s.name),
-      })
-    : {};
-  // Daytona's provider is built from `piExtEnv` rather than the local daemon env. Keep the
-  // transcript location in both environment slices so Pi and pi-acp see the same durable path
-  // regardless of provider.
-  if (piSessionDir) piExtEnv.PI_CODING_AGENT_SESSION_DIR = piSessionDir;
-  configurePiSkillSnapshot(piSkillSnapshot, piExtEnv);
-  Object.assign(env, piExtEnv); // local daemon inherits it; daytona gets it via envVars
-  logger(
-    `tools=${plan.toolSpecs.length} executableTools=${plan.executableToolSpecs.length} ` +
-      `piPublicTools=${piExtEnv.AGENTA_AGENT_TOOLS_PUBLIC_SPECS ? "yes" : "no"}`,
-  );
-  if (!plan.isPi && plan.isDaytona) {
-    const omittedClientTools = plan.toolSpecs
-      .filter((spec) => spec.kind === "client")
-      .map((spec) => spec.name);
-    if (omittedClientTools.length > 0) {
-      logger(
-        `omitting client tools from Daytona stdio MCP shim: ${omittedClientTools.join(", ")}`,
-      );
-    }
-  }
-  // undefined is fine: the local provider runs its own resolution and errors clearly.
-  const binaryPath = (deps.resolveDaemonBinary ?? resolveDaemonBinary)();
-  const localPiAssets = prepareLocalPiAssets({ plan, env, log: logger });
-  let runAgentDir = localPiAssets.dir;
-  // Fail closed (Decision 2): when the policy could gate a Pi built-in tool but the permission
-  // extension did not install, the run must stop rather than run those tools unprotected. Recorded
-  // here (the install ran above) and thrown inside the try below so the engine's own catch turns it
-  // into `{ ok: false, error }` and a visible error frame. `builtinGatingActive` false means
-  // allow-everything, where the extension is not needed and a failed install is harmless.
-  const localBuiltinGatingUnenforceable =
-    plan.isPi &&
-    !plan.isDaytona &&
-    plan.builtinGatingActive &&
-    !localPiAssets.extensionInstalled;
-
-  // A local Claude subscription run reads and writes the operator's read-write mounted login
-  // DIRECTLY: `buildDaemonEnv` already carried `CLAUDE_CONFIG_DIR` (the mount) into the daemon env,
-  // and there is deliberately no per-run copy. Claude refreshes its OAuth token mid-run and writes
-  // it back to its config dir; copying that dir per run would discard the refresh, so the next run
-  // would fail as soon as the provider rotated the refresh token. The harness owns its own token
-  // lifecycle, exactly like a normal local install (interface.md section 6). buildRunPlan already
-  // rejected a runtime_provided Claude run with no configured CLAUDE_CONFIG_DIR.
-
-  logger(`harness=${plan.harness} sandbox=${plan.sandboxId} cwd=${plan.cwd}`);
-
-  // The resolved model ref as it reaches the runner (key NAMES only, never values) — the one
-  // line that answers "what model/provider/deployment/credential did this run actually use".
-  logger(
-    `resolved model=${request.model ?? "<none>"} provider=${request.provider ?? "<none>"} ` +
-      `deployment=${request.deployment ?? "<none>"} ` +
-      `connection=${request.connection ? `${request.connection.mode}:${request.connection.slug ?? "-"}` : "<none>"} ` +
-      `secretKeys=[${Object.keys(request.secrets ?? {}).join(",")}]`,
-  );
-
-  // The shared client-tool relay reference (the deferred ref baked into the MCP server reads it;
-  // each turn's `runTurn` sets `.current`). A `tools/call` can only arrive during a prompt —
-  // long after the relay is wired — so the server captures this reference and it resolves to the
-  // real relay before any call lands.
-  const clientToolRelayRef: { current?: ClientToolRelay } = {};
-  const deferredClientToolRelay: ClientToolRelay = {
-    onClientTool: (req) =>
-      clientToolRelayRef.current
-        ? clientToolRelayRef.current.onClientTool(req)
-        : Promise.resolve("deny" as ClientToolOutcome),
-    onPause: (req) => clientToolRelayRef.current?.onPause?.(req),
-  };
-
-  // Aborts any in-flight loopback `tools/call` (a paused Claude client tool) on pause/teardown,
-  // so its handler is torn down deterministically and cannot write a result after the turn ends.
-  const mcpAbort = new AbortController();
-
-  const environment: SessionEnvironment = {
-    plan,
+  const setup = await prepareEnvironmentSetup(request, deps, presignedMount);
+  if (!setup.ok) return setup;
+  const {
+    acquireStartedAt,
+    agentMountDir,
+    artifactId,
+    binaryPath,
+    deferredClientToolRelay,
+    env,
+    environment,
+    localBuiltinGatingUnenforceable,
     logger,
-    deps,
-    sandbox: undefined,
-    session: undefined,
-    sessionId: resolveRunSessionId(request, ""),
-    model: undefined,
-    capabilities: {},
-    strictModel,
-    toolCallIndex: createToolCallCorrelationIndex(),
-    clientToolRelayRef,
     mcpAbort,
-    runAgentDir,
-    otlpAuthFilePath,
-    mountCreds,
-    agentMountCreds,
-    mountProjectId: mountCreds?.projectId,
-    projectScopeId: projectScopeFor(request, mountCreds?.projectId)?.id,
-    loadedFromContinuity: false,
-    resumable: false,
-    continuityTurnIndex: undefined,
-    sessionDestroyRequested: false,
-    mountedCwd: undefined,
-    agentMountedPath: undefined,
-    durableCwdSafeToDelete: true,
-    // Local runs get a plain rmSync cleanup for the throwaway cwd; Daytona has none on this host.
-    workspace: plan.isDaytona
-      ? undefined
-      : {
-          cleanup: async () =>
-            rmSync(plan.cwd, { recursive: true, force: true }),
-        },
-    runtimeRemount: undefined,
-    closeToolMcp: undefined,
-    currentTurn: undefined,
-    lastTurnToolCallIds: [],
-    parkedApproval: undefined,
-    approvalGateCount: 0,
-    destroyed: false,
-    destroy: async () => {},
-    clearTurn: () => {},
-  };
-
-  environment.clearTurn = () => {
-    environment.currentTurn = undefined;
-  };
+    piExtEnv,
+    piSessionDir,
+    piSkillSnapshot,
+    plan,
+    runCred,
+    sessionForMount,
+    signAgentMount,
+    signMount,
+    strictModel,
+    timingLog,
+  } = setup;
+  let runAgentDir = setup.runAgentDir;
 
   // The one complete, idempotent teardown — the same steps the old per-run `finally` ran, in the
   // same order. Every resource is null-checked, so it is safe after a partial acquire and safe to
